Refactored and optimized for the server
