+++ core/settings.py
@@ -72,7 +72,6 @@ SHARED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'django_extensions',
 
     'django_ckeditor_5',
     'rest_framework',
@@ -166,7 +165,6 @@ DATABASES = {
         'PORT': os.getenv("DB_PORT")
     }
 }
-print("databasename", os.getenv("DB_NAME") )
 
 DATABASE_ROUTERS = (
     'django_tenants.routers.TenantSyncRouter',
