+++ registration/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
-from companies.views import LoginView
-from .views import TenantRegistrationViewSet
+from .views import TenantRegistrationViewSet, LoginView
+
 
 urlpatterns = [
     path('register/', TenantRegistrationViewSet.as_view({'post': 'create'}), name='register'),
